adding logic for budget, and made changes to get info from DB.
