+++ src/main/java/com/tgj/eventaid/controllers/EventsController.java
@@ -1,15 +1,12 @@
 package com.tgj.eventaid.controllers;
 
 import com.tgj.eventaid.models.Artist;
-import com.tgj.eventaid.models.Budget;
 import com.tgj.eventaid.models.Event;
 import com.tgj.eventaid.models.User;
 import com.tgj.eventaid.repositories.ArtistsRepository;
-import com.tgj.eventaid.repositories.BudgetRepository;
 import com.tgj.eventaid.repositories.EventsRepository;
 import com.tgj.eventaid.repositories.UserRepository;
 import org.springframework.beans.propertyeditors.CustomDateEditor;
-import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -17,10 +14,8 @@ import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.text.SimpleDateFormat;
 import java.util.Date;
-import java.util.List;
 
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -30,13 +25,11 @@ import org.springframework.web.bind.annotation.PostMapping;
 public class EventsController {
 
     public EventsRepository eventsRepository;
-    public BudgetRepository budgetRepository;
     public UserRepository userRepository;
     public ArtistsRepository artistsRepository;
 
-    public EventsController(EventsRepository eventsRepository, BudgetRepository budgetRepository, UserRepository userRepository, ArtistsRepository artistsRepository) {
+    public EventsController(EventsRepository eventsRepository, UserRepository userRepository, ArtistsRepository artistsRepository) {
         this.eventsRepository = eventsRepository;
-        this.budgetRepository = budgetRepository;
         this.userRepository = userRepository;
         this.artistsRepository = artistsRepository;
     }
@@ -69,7 +62,6 @@ public class EventsController {
     @PostMapping("/events/create")
     public String saveEvent(@ModelAttribute Event event,
                             @RequestParam ("upload") String picture,
-                            @ModelAttribute Budget budget,
                             @RequestParam ("event_budget") BigDecimal event_budget,
                             @RequestParam ("target_profit") BigDecimal target_profit,
                             @RequestParam ("target_spending") BigDecimal target_spending,
@@ -85,12 +77,6 @@ public class EventsController {
         event.setUser(user);
         event.setOwner(user);
         eventsRepository.save(event);
-        //saving info to budget tables
-        budget.setEvent(event);
-        budget.setEvent_budget(event_budget);
-        budget.setTarget_profit(target_profit);
-        budget.setTarget_spending(target_spending);
-        budgetRepository.save(budget);
         //saving info to artists table
         artist.setEvent(event);
         artist.setName(artist_name);
